Refactor demo code to reduce distraction from Zustand usage
