+++ examples/src/components/CodePreview.js
@@ -1,11 +1,11 @@
 import React from 'react'
 import Highlight, { defaultProps } from 'prism-react-renderer'
-import { CopyButton } from './CopyButton'
+import CopyButton from './CopyButton'
 import 'prismjs'
 import 'prismjs/components/prism-jsx.min'
 import 'prismjs/themes/prism-okaidia.css'
 
-function CodePreview({ code, ...props }) {
+export default function CodePreview({ code, ...props }) {
   return (
     <Highlight {...defaultProps} className="language-jsx" code={code} language="jsx" theme={undefined}>
       {({ className, style, tokens, getLineProps, getTokenProps }) => (
@@ -25,4 +25,3 @@ function CodePreview({ code, ...props }) {
     </Highlight>
   )
 }
-export default CodePreview
