+++ examples/src/components/CopyButton.js
@@ -5,7 +5,7 @@ import { copyToClipboard } from '../utils/copy-to-clipboard'
 Isolated logic for the entire copy functionality instead 
 of a separate button component and with the added utility
 */
-export function CopyButton({ code, ...props }) {
+export default function CopyButton({ code, ...props }) {
   const [isCopied, setIsCopied] = useState(false)
 
   const handleCopy = () => {
@@ -31,7 +31,8 @@ export function CopyButton({ code, ...props }) {
               strokeWidth={2}
               strokeLinecap="round"
               strokeLinejoin="round"
-              {...props}>
+              {...props}
+            >
               <rect x={9} y={9} width={13} height={13} rx={2} ry={2} />
               <path d="M5 15H4a2 2 0 01-2-2V4a2 2 0 012-2h9a2 2 0 012 2v1" />
             </svg>
