+++ examples/src/components/Details.js
@@ -1,0 +1,18 @@
+import React from 'react'
+
+export default function () {
+  return (
+    <>
+      <a href="https://github.com/drcmda/zustand" className="top-right" children="Github" />
+      <div className="bottom">
+        <a href="https://codesandbox.io/s/xgjtc" className="bottom-right" children="<Source />" />
+        <a
+          href="https://www.instagram.com/tina.henschel/"
+          className="bottom-left"
+          children="Illustrations @ Tina Henschel"
+        />
+      </div>
+      <span className="header-left">Zustand</span>
+    </>
+  )
+}
